fix: resolve all flutter analyze errors and warnings (enum caps, unnecessary_cast, avoid_print)
Agent-Logs-Url: https://github.com/chjsoliveira/maps_place_picker/sessions/39f79a91-b6df-4e22-a367-08cab1807cf6

+++ lib/src/components/prediction_tile.dart
@@ -33,8 +33,8 @@ class PredictionTile extends StatelessWidget {
 
     if (prediction.matchedSubstrings.isNotEmpty) {
       final MatchedSubstring matchedSubString = prediction.matchedSubstrings[0];
-      final int offset = matchedSubString.offset as int;
-      final int length = matchedSubString.length as int;
+      final int offset = matchedSubString.offset;
+      final int length = matchedSubString.length;
       final int end = offset + length;
 
       // Guard against out-of-bounds offsets returned by the API.
